pyproject.toml
[project]
name = "behavior-stream-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Fábio Henrique",email = "psifabiohenrique@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "behavior_stream_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
